fix: address issue 579 mapping dialog state effects
